+++ gradle-bridge/Cargo.toml
@@ -11,7 +11,7 @@ name = "gradle-lsp-bridge"
 path = "src/main.rs"
 
 [dependencies]
-proxy-common.workspace = true
+proxy-common = { workspace = true, features = ["tokio"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 prost = "0.14"
