+++ proxy-common/src/lib.rs
@@ -11,6 +11,8 @@ pub mod lsp;
 pub mod platform;
 pub mod uri;
 
+#[cfg(feature = "tokio")]
+pub use lsp::AsyncLspReader;
 pub use lsp::{
     contains_subslice, encode_lsp, lsp_body, parse_content_length, parse_lsp_content, raw_has_id,
     write_raw, write_to_stdout, LspReader, CONTENT_LENGTH, HEADER_SEP,
